+++ FRONT/src/components/NavbarComponent.vue
@@ -8,12 +8,13 @@
 
       <v-spacer></v-spacer>
 
-      <RouterLink to="/favorites">
+      <RouterLink to="/favs">
         <v-btn variant="text" icon="mdi-heart"></v-btn>
       </RouterLink>
 
-      <v-btn variant="text" icon="mdi-pencil" @click.stop="overlay = !overlay"></v-btn>
-
+      <RouterLink to="/products">
+       <v-btn variant="text" icon="mdi-pencil" @click.stop="overlay = !overlay"></v-btn>
+      </RouterLink>
       <v-btn variant="text" icon="mdi-dots-vertical"></v-btn>
     </v-app-bar>
 
@@ -22,17 +23,13 @@
         <RouterLink
           class="itemList"
           v-for="item in items"
-          :key="item.title"
+          :key="item.name"
           :to="item.route"
         >
           {{ item.title }}
         </RouterLink>
       </v-list>
     </v-navigation-drawer>
-
-    <!-- <div class="image-container">
-      <img class="grayscale" src="/public/tshirt_back.jpg" alt="fabrica de camisetas" />
-    </div> -->
 
     <v-overlay v-model="overlay" class="d-flex align-center justify-center" scrim="#000" @click:outside="overlay = false">
       <div class="my-overlay-content" style="max-height: 80vh; overflow-y: auto; background-color: white; z-index: 2001;">
@@ -46,6 +43,10 @@
 import { ref, watch, onMounted } from 'vue';
 import { RouterLink } from 'vue-router';
 import FormComponent from '@/components/FormComponent.vue';
+import Team from '@/views/Team.vue';
+
+
+
 
 const drawer = ref(false);
 const group = ref(null);
@@ -53,24 +54,20 @@ const overlay = ref(false);
 
 const items = [
   {
-    title: 'Home',
+    name: 'Home',
     route: '/',
   },
   {
-    title: 'Team',
+    name: 'Team',
     route: '/team',
   },
   {
-    title: 'Services',
-    route: '/services',
-  },
-  {
-    title: 'Contact Us',
-    route: '/contact',
-  },
-  {
-    title: 'Login',
+    name: 'Login',
     route: '/login',
+  },
+  {
+    name: 'SignUp',
+    route: '/SignUp',
   },
 ];
